requirements.txt
numpy==1.24.2
segment_anything==1.0
transformers==4.27.4
six==1.16.0
kornia==0.5.0
imgaug==0.4.0
tabulate==0.8.10
torch==1.13.1+cu117
easydict==1.9
iopath==0.1.9
pandas==2.0.0
torchvision==0.14.1+cu117
tqdm==4.61.2
webdataset==0.2.33
pytorch_lightning==2.0.1
matplotlib==3.7.1
Pygments==2.14.0
diffusers==0.17.0.dev0
imutils==0.5.4
opencv_python_headless==4.6.0.66
omegaconf==2.2.3
Shapely==1.8.5.post1
fvcore==0.1.5.post20221122
timm==0.6.13
scikit_image==0.19.3
joblib==1.1.0
albumentations==0.5.2
imageio==2.27.0
regex==2022.10.31
caffe2==0.8.1
cityscapesscripts==2.2.2
dataclasses==0.8
einops==0.6.1
fairscale==0.4.13
fastapi==0.95.2
ftfy==6.1.1
gradio==3.32.0
memory_profiler==0.61.0
mmcv==2.0.0
mmdet==3.0.0
onnx==1.14.0
packaging==23.1
Pillow==9.5.0
PyYAML==6.0
scikit_learn==1.2.2
scipy==1.10.1
skimage==0.0
sync_batchnorm==0.0.1
termcolor==2.3.0
wandb==0.15.3
